// src-shared/interfaces/theme.interface.ts
/* eslint-disable @typescript-eslint/naming-convention */

export interface ThemeColors {
	'primary': string
	'primary-focus': string
	'primary-content': string
	'secondary': string
	'secondary-focus': string
	'secondary-content': string
	'accent': string
	'accent-focus': string
	'accent-content': string
	'neutral': string
	'neutral-focus': string
	'neutral-content': string
	'base-100': string
	'base-200': string
	'base-300': string
	'base-content': string
	'info': string
	'info-content': string
	'success': string
	'success-content': string
	'warning': string
	'warning-content': string
	'error': string
	'error-content': string
}